Hoist components that will be shared
* We will need ModalWizard on both the Overview and Mappings screens
* This empty state is implemented specifically to launch the wizards, so
  rename to reflect this. We will need this component on both the
  Overview and Mappings screens
