+++ app/javascript/react/screens/App/App.scss
@@ -2,3 +2,5 @@
 @import './Overview/Overview.scss';
 @import './Plan/Plan.scss';
 @import './Mappings/Mappings.scss';
+@import './common/ModalWizard/ModalWizard.scss';
+@import './common/ShowWizardEmptyState/ShowWizardEmptyState.scss';
