+++ app/javascript/react/screens/App/Overview/components/InfrastructureMappingsList/__test__/InfrastructureMappingList.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 
 import InfrastructureMappingsList from '../InfrastructureMappingsList';
-import OverviewEmptyState from '../../OverviewEmptyState/OverviewEmptyState';
+import ShowWizardEmptyState from '../../../../common/ShowWizardEmptyState/ShowWizardEmptyState';
 
 let createInfraMappingClick;
 beforeEach(() => {
@@ -17,7 +17,7 @@ test('it shows the create infra mapping empty state when there are no transforma
       createInfraMappingClick={createInfraMappingClick}
     />
   );
-  expect(wrapper.find(OverviewEmptyState).props().buttonText).toBe('Create Infrastructure Mapping');
+  expect(wrapper.find(ShowWizardEmptyState).props().buttonText).toBe('Create Infrastructure Mapping');
 });
 
 test('it shows the error empty state when there is an error loading mappings', () => {
@@ -28,5 +28,5 @@ test('it shows the error empty state when there is an error loading mappings', (
       error
     />
   );
-  expect(wrapper.find(OverviewEmptyState).props().iconName).toBe('error-circle-o');
+  expect(wrapper.find(ShowWizardEmptyState).props().iconName).toBe('error-circle-o');
 });
