+++ app/javascript/react/screens/App/common/ModalWizard/ModalWizard.js
