+++ app/javascript/react/screens/App/common/ModalWizard/ModalWizardBody.js
