+++ app/javascript/react/screens/App/common/ModalWizard/ModalWizardStateProvider.js
