+++ app/javascript/react/screens/App/common/ModalWizard/index.js
