+++ app/javascript/react/screens/App/common/ShowWizardEmptyState/ShowWizardEmptyState.scss
